Fake RADIO_STATUS, doc updates and repo shuffle.
Radio status message is now being packed, so the empty buffers are now
being populated with the sought-after array of bytes to be passed into
the XBee tx buffer.  Began separating different portions of the
repository into different directories starting with USB device rules
files.  After the code and documentation is polished off a little more
the same will be done for these.